feat(all): convert to TypeScript
