--- doc/core-js.d.ts
@@ -1,4 +1,0 @@
-declare module 'core-js' {
-        var coreJs;
-        export default coreJs;
-}
