+++ jsconfig.json
@@ -1,0 +1,6 @@
+{
+	"compilerOptions": {
+		"target": "ES6",
+		"module": "system"
+	}
+}
