+++ package.json
@@ -49,8 +49,10 @@
     "gulp-bump": "^0.3.1",
     "gulp-eslint": "^1.0.0",
     "gulp-rename": "^1.2.2",
+    "gulp-sourcemaps": "^1.6.0",
     "gulp-typedoc": "^1.2.1",
     "gulp-typedoc-extractor": "^0.0.8",
+    "gulp-typescript": "^2.9.2",
     "jasmine-core": "^2.1.3",
     "karma": "^0.13.15",
     "karma-babel-preprocessor": "^5.2.2",
@@ -58,9 +60,11 @@
     "karma-coverage": "^0.3.1",
     "karma-jasmine": "^0.3.5",
     "karma-jspm": "^2.0.1",
+    "merge2": "^0.3.6",
     "object.assign": "^1.0.3",
     "require-dir": "^0.1.0",
     "run-sequence": "^1.0.2",
+    "typescript": "^1.6.2",
     "vinyl": "^0.5.1",
     "vinyl-paths": "^1.0.0",
     "yargs": "^2.1.1"
