+++ typings/Interfaces.d.ts
@@ -1,0 +1,3 @@
+interface Constructor<T> {
+    new (...args: any[]): T;
+}
